+++ src/app/(tabs)/index.jsx
@@ -1,6 +1,4 @@
-import { Text } from "react-native"
 import { observer } from "mobx-react-lite"
-import { reactiveModel } from "src/bootstrapping"
 import { Request } from "src/presenters/requestPresenter"
 
 export default observer(function IndexPage() {
