+++ src/app/index.jsx
@@ -1,7 +1,9 @@
 import { observer } from "mobx-react-lite"
-import { reactiveModel } from "src/bootstrapping"
 import { Login } from "src/presenters/loginPresenter"
+import { persistModel } from "./firebasemodel"
+import { model } from "./model"
 
 export default observer(function IndexPage() {
   return <Login />
 })
+persistModel(model);
